refactor: move all constants to a single folder
