--- src/components/API/data.tsx
@@ -1,8 +1,0 @@
-const ROOT_ENDPOINT: string = 'https://api.punkapi.com/v2/';
-const BEERS: string = 'beers?';
-const PAGE_NUMBER: string = 'page=';
-const PER_PAGE: string = '&per_page=';
-const BEER_NAME: string = '&beer_name=';
-const SINGLE_BEER: string = 'https://api.punkapi.com/v2/beers/';
-
-export { ROOT_ENDPOINT, BEERS, PAGE_NUMBER, PER_PAGE, BEER_NAME, SINGLE_BEER };
