--- src/components/Paginator/data.ts
@@ -1,3 +1,0 @@
-const ENTER_MESSAGE = 'Please, set a page number and hit "Enter"';
-
-export default ENTER_MESSAGE;
